refactor: type 폴더 구조화 및 중복타입 제거 완료(채팅 타입)
