+++ src/app/(pages)/(chat)/_components/oneTimeClub/OneTimeClubChatRoomMeetingPlace.tsx
@@ -1,8 +1,8 @@
 // 모임 채팅 생성후 모임장 채팅에 입장 시키기
 
-import { OneTimeChatRoom } from "../../types/chat";
+import { EggPopChatRoom } from "@/types/eggpopchat.types";
 
-export async function OneTimeClubChatRoomMeetingPlace(OneTimeChatRoom: OneTimeChatRoom) {
+export async function OneTimeClubChatRoomMeetingPlace(OneTimeChatRoom: EggPopChatRoom) {
   try {
     await fetch("/api/oneTimeChatRoomMeetingPlace", {
       method: "POST",
