--- src/app/(pages)/(chat)/types/r_c_n_chatting_room.ts
@@ -1,4 +1,0 @@
-export type chattingRoom = {
-  r_c_n_chatting_room_id: number;
-  r_c_n_chatting_room_name: string;
-};
