+++ src/types/user.types.ts
@@ -1,0 +1,6 @@
+// 유저 타입
+export type User = {
+  user_id: string;
+  user_name: string;
+  user_profile_img: string;
+};
